Create 'Record' classes to store data pulled from Elements
Why these changes are being introduced:
* The purpose of these 'Record' classes is to create an abstraction
containing the raw data pulled from Elements. These abstractions are
different from the data models defined in 'pers.database', which
represent the final representation of an author, publication, etc.
in the PERS database. These 'Record' classes also serve as a way to
distinguish the differing (a) API calls and (b) data processing steps
for a given record *before* they are upserted into the database.

How this addresses that need:
* Create 'AuthorRecord'
* Create 'PublicationRecord'